[core] refactored structure (DI and AOP = > /compoonent)
