+++ vertigo-core/src/main/java/io/vertigo/core/component/di/DIAnnotationUtil.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.vertigo.core.di;
+package io.vertigo.core.component.di;
 
 import io.vertigo.lang.Assertion;
 import io.vertigo.util.StringUtil;
