+++ vertigo-core/src/test/java/io/vertigo/AbstractTestCaseJU4.java
@@ -21,8 +21,8 @@ package io.vertigo;
 import io.vertigo.boot.xml.XMLAppConfigBuilder;
 import io.vertigo.core.App;
 import io.vertigo.core.Home;
+import io.vertigo.core.component.di.injector.Injector;
 import io.vertigo.core.config.AppConfig;
-import io.vertigo.core.di.injector.Injector;
 import io.vertigo.core.spaces.component.ComponentInfo;
 import io.vertigo.lang.Component;
 import io.vertigo.lang.Describable;
