+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/task/TaskManagerImpl.java
@@ -19,7 +19,7 @@
 package io.vertigo.dynamo.impl.task;
 
 import io.vertigo.core.Home;
-import io.vertigo.core.di.injector.Injector;
+import io.vertigo.core.component.di.injector.Injector;
 import io.vertigo.dynamo.impl.task.listener.TaskListener;
 import io.vertigo.dynamo.impl.task.listener.TaskListenerImpl;
 import io.vertigo.dynamo.task.TaskManager;
