+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/transaction/VTransactionAspect.java
@@ -18,8 +18,8 @@
  */
 package io.vertigo.dynamo.impl.transaction;
 
-import io.vertigo.core.aop.Aspect;
-import io.vertigo.core.aop.AspectMethodInvocation;
+import io.vertigo.core.component.aop.Aspect;
+import io.vertigo.core.component.aop.AspectMethodInvocation;
 import io.vertigo.dynamo.transaction.VTransactionManager;
 import io.vertigo.dynamo.transaction.VTransactionWritable;
 import io.vertigo.dynamo.transaction.Transactional;
