+++ vertigo-tempo-impl/src/main/java/io/vertigo/tempo/impl/job/JobManagerImpl.java
@@ -20,7 +20,7 @@ package io.vertigo.tempo.impl.job;
 
 import io.vertigo.commons.analytics.AnalyticsManager;
 import io.vertigo.core.Home;
-import io.vertigo.core.di.injector.Injector;
+import io.vertigo.core.component.di.injector.Injector;
 import io.vertigo.lang.Assertion;
 import io.vertigo.tempo.job.JobManager;
 import io.vertigo.tempo.job.metamodel.JobDefinition;
