+++ vertigo-vega-impl/src/main/java/io/vertigo/vega/plugins/rest/handler/JsonConverterRestHandlerPlugin.java
@@ -19,7 +19,7 @@
 package io.vertigo.vega.plugins.rest.handler;
 
 import io.vertigo.core.Home;
-import io.vertigo.core.di.injector.Injector;
+import io.vertigo.core.component.di.injector.Injector;
 import io.vertigo.dynamo.domain.model.DtList;
 import io.vertigo.dynamo.domain.model.DtObject;
 import io.vertigo.lang.Assertion;
